Move dialogs to separate class. Add confirm dialog.

+++ src/main/java/net/techmayhem/clocktrack/screens/people/PeopleOverview.java
@@ -4,15 +4,14 @@ import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.geometry.Insets;
 import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.*;
+import javafx.scene.control.Button;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
-import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
 import net.techmayhem.clocktrack.Database;
+import net.techmayhem.clocktrack.dialog.Dialogs;
 import net.techmayhem.clocktrack.models.FromDb;
 import net.techmayhem.clocktrack.models.Person;
 import net.techmayhem.clocktrack.screens.Overview;
@@ -80,38 +79,7 @@ public class PeopleOverview extends Overview {
     }
 
     private void spawnNewPersonDialog() {
-        spawnTextInputDialog("Add Person", "", "Add", this::createPerson);
-    }
-
-    private void spawnTextInputDialog(String title, String defaultValue, String accept, Consumer<String> callback) {
-        Stage dialog = new Stage();
-        dialog.setTitle(title);
-        dialog.setResizable(false);
-        dialog.initModality(Modality.APPLICATION_MODAL);
-        dialog.initOwner(root.getScene().getWindow());
-        VBox dialogVbox = new VBox(20);
-        dialogVbox.setPadding(new Insets(20));
-        Label label = new Label("Enter the name of the person");
-        TextField textInput = new TextField(defaultValue);
-        HBox hBox = new HBox();
-        Button acceptButton = new Button(accept);
-        acceptButton.setDefaultButton(true);
-        acceptButton.setOnAction(_ -> {
-            callback.accept(textInput.getText());
-            dialog.close();
-        });
-        acceptButton.setDisable(true);
-        textInput.setOnKeyTyped(_ -> acceptButton.setDisable(textInput.getText().isEmpty()));
-        Region spacer = new Region();
-        HBox.setHgrow(spacer, Priority.ALWAYS);
-        Button cancelButton = new Button("Cancel");
-        cancelButton.setCancelButton(true);
-        cancelButton.setOnAction(_ -> dialog.close());
-        hBox.getChildren().addAll(acceptButton, spacer, cancelButton);
-        dialogVbox.getChildren().addAll(label, textInput, hBox);
-        Scene dialogScene = new Scene(dialogVbox);
-        dialog.setScene(dialogScene);
-        dialog.show();
+        Dialogs.showTextDialog("Add Person", "Enter the name of the person to add", "", "Add", "Cancel", this::createPerson);
     }
 
     private void createPerson(String name) {
@@ -121,8 +89,10 @@ public class PeopleOverview extends Overview {
 
     private void deleteSelectedPerson() {
         FromDb<Person> selectedPerson = table.getSelectionModel().getSelectedItem();
-        Database.getInstance().deletePerson(selectedPerson.id());
-        updateTable();
+        Dialogs.showConfirmDialog("Confirm deletion", "Do you really want to delete " + selectedPerson.model().getName() + "?", "Delete", "Cancel", () -> {
+            Database.getInstance().deletePerson(selectedPerson.id());
+            updateTable();
+        });
     }
 
     @Override
@@ -132,7 +102,7 @@ public class PeopleOverview extends Overview {
 
     private void spawnEditPersonDialog() {
         FromDb<Person> selectedPerson = table.getSelectionModel().getSelectedItem();
-        spawnTextInputDialog("Edit Person", selectedPerson.model().getName(), "Rename", s -> updateName(selectedPerson, s));
+        Dialogs.showTextDialog("Edit Person", "Enter the new name of the person", selectedPerson.model().getName(), "Rename", "Cancel", s -> updateName(selectedPerson, s));
     }
 
     private void updateName(FromDb<Person> person, String newName) {
